% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiregulon_igraph.R
\name{add_centrality_degree}
\alias{add_centrality_degree}
\title{Calculate degree centrality}
\usage{
add_centrality_degree(graph)
}
\arguments{
\item{graph}{an igraph object}
}
\value{
an igraph object with attribute \code{centrality} added to vertices
}
\description{
Calculate degree centrality for each vertex
}
