% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archr2MAE.R
\name{archr2MAE}
\alias{archr2MAE}
\title{Create a MultiAssayExperiment from an ArchR project}
\usage{
archr2MAE(archrDir, defaultMatrix = "TileMatrix")
}
\arguments{
\item{archrDir}{character string specifying the ArchR project directory,
which must contain a file called Save-ArchR-Project.rds}

\item{defaultMatrix}{character string specifying the name of the matrix that
unmapped embeddings and reduced dimension matrices should be associated with}
}
\value{
A \linkS4class{MultiAssayExperiment}
}
\description{
Import the matrices from a ArchR project to into a MultiAssayExperiment,
where each ArchR matrix is a SingleCellExperiment.
Metadata is not added to the MultiAssayExperiment or Experiments.
}
\author{
Natalie Fox
}
