% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiregulon_igraph.R
\name{build_difference_graph}
\alias{build_difference_graph}
\title{Build a graph difference}
\usage{
build_difference_graph(
  graph_obj_1,
  graph_obj_2,
  weighted = TRUE,
  abs_diff = TRUE
)
}
\arguments{
\item{graph_obj_1}{an igraph object from which \code{graph_obj_2} will be
subtracted}

\item{graph_obj_2}{an igraph object used for being subtracted from \code(graph_obj_1)}

\item{weighted}{a logical indicating whether weighted graphs are used}

\item{abs_diff}{a logical indicating whether absulute difference in the number
edges or their weights will be calculated}
}
\value{
an igraph object
}
\description{
Build a graph based on the edge difference of two input graphs
}
\details{
Function building a graph difference by subtracting the edges of \code{graph_obj_2}
from those of the \code{graph_obj_1}. If \code{weighted} is set to \code{TRUE} then for each
ordered pair of vertices (nodes) the difference in number of edges between \code{graph_obj_1}
and \code{graph_obj_1} is calculated. The result is used to set the number of
corresponding edges in output graph. Note that unless \code{abs_diff} is set to
\code{TRUE} any non-positive difference will translate into lack of the edges
for a corresponding ordered pair of vertices in the output graph (equivalent
to 0 value in the respective position in adjacency matrix). In case of
weighted graphs, the weight of the output graph is calculated as a difference
of the corresponding weights between input graphs.
}
