% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiregulon_igraph.R
\name{build_graph}
\alias{build_graph}
\title{Building igraph directed graph object based on the output of the getRegulon function}
\usage{
build_graph(regulon, mode = "tripartite", weights = "corr")
}
\arguments{
\item{regulon}{An object returned by the getRegulon or addWeights function}

\item{mode}{A character specifying whch type of graph will be built. In 'tg' mode
a bipartite graph is built connecting transcription factors directly to target genes
and ignoring information about mediating regulatory elements; in 'pairs' mode
transcription factors are connected to unique target gene-regulatory element pairs;
in tripartite mode the network is build of three type of vertices (nodes):
trascription factors, regulatory elements and target genes; here the path from
target gene to regulatory element always contains a regulatory elements}

\item{weights}{A character specifying which variable should be used to assign
weights to edges. If set to 'NA' then unweighted graph is built.}
}
\value{
A regulatory network graph
}
\description{
Building igraph directed graph object based on the output of the getRegulon function
}
