% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonSimilarities.R
\name{calculateJaccardSimilarity}
\alias{calculateJaccardSimilarity}
\title{Calculate Jaccard Similarity between regulons of all transcription factors}
\usage{
calculateJaccardSimilarity(graph)
}
\arguments{
\item{graph}{a igraph object from \code{buildGraph} or \code{buildDiffGraph}}
}
\value{
A matrix with Jaccard similarity between all pairs of transcription factors.
}
\description{
Calculate Jaccard Similarity between regulons of all transcription factors
}
