% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{enrichPlot}
\alias{enrichPlot}
\title{A function to plot results of regulonEnrich}
\usage{
enrichPlot(results, top = 15, ncol = 3, combine = TRUE)
}
\arguments{
\item{results}{regulonEnrich}

\item{top}{the number of pathways to plot ranked by significance. Default is 15.}

\item{ncol}{number of columns in the combined plot, if combine == TRUE. Default ncol is 3.}

\item{combine}{whether to combine and visualize the plots in one panel or not}
}
\value{
a combined ggplot object or a list of ggplots if combine == FALSE
}
\description{
A function to plot results of regulonEnrich
}
\examples{
enrichplot(results=enrichment_results )
}
