% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiregulon-package.R
\docType{package}
\name{epiregulon-package}
\alias{epiregulon}
\alias{epiregulon-package}
\title{epiregulon: Gene regulatory network inference from single cell epigenomic data}
\description{
Gene regulatory networks model the underlying gene regulation hierarchies that drive gene expression and observed phenotypes. Epiregulon infers TF activity in single cells by constructing a gene regulatory network (regulons). This is achieved through integration of scATAC-seq and scRNA-seq data and incorporation of public bulk TF ChIP-seq data. Links between regulatory elements and their target genes are established by computing correlations between chromatin accessibility and gene expressions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xiaosaiyao/epiregulon/}
  \item Report bugs at \url{https://github.com/xiaosaiyao/epiregulon/issues}
}

}
\author{
\strong{Maintainer}: Xiaosai Yao \email{xiaosai.yao@gmail.com} (\href{https://orcid.org/0000-0001-9729-0726}{ORCID})

Authors:
\itemize{
  \item Xiaosai Yao \email{xiaosai.yao@gmail.com} (\href{https://orcid.org/0000-0001-9729-0726}{ORCID})
  \item Tomasz Włodarczyk \email{tomwlo@gmail.com} (\href{https://orcid.org/0000-0003-1554-9699}{ORCID})
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}
  \item Shang-Yang Chen \email{sychen9584@gmail.com}
}

}
\keyword{internal}
