% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers.R
\name{findMarkersSimple}
\alias{findMarkersSimple}
\title{Find Differential Genes}
\usage{
findMarkersSimple(
  x,
  groups,
  direction = c("any", "up", "down"),
  pval.type = c("any", "some", "all"),
  min.prop = NULL,
  assay.type = "logcounts",
  combined = TRUE
)
}
\arguments{
\item{x}{A matrix or SummarizedExperiment object}

\item{groups}{A vector of length equal to the number of cells, indicating the group of each cell}

\item{direction}{A string specifying the direction of log-fold changes to be considered in the alternative hypothesis.
See \code{scran::pairwiseTTests}}

\item{pval.type}{A string specifying how p-values are to be combined across pairwise comparisons for a given group.
See \code{scran::combineMarkers}}

\item{min.prop}{Numeric scalar specifying the minimum proportion of significant comparisons per gene.
Defaults to 0.5 when pval.type="some", otherwise defaults to zero.  See \code{scran::combineMarkers}}

\item{assay.type}{A string specifying which assay values to use. Default is \code{logcounts}}

\item{combined}{logical indicating pairwise p-values should be combined}
}
\value{
If combined is FALSE, this function returns a list of 4 elements:
\itemize{
\item \code{log.p.value}: A list of data.frame, of length equal to the number of groups, indicating the
log p-values of each group member vs versus all other group members
\item \code{log.FDR}: A list of data.frame, of length equal to the number of groups, indicating the
log FDR of each group member vs versus all other group members
\item \code{logFC}: A list of data.frame, of length equal to the number of groups, indicating the
log fold changes of each group member vs versus all other group members
\item \code{all.groups}: a vector corresponding to the unique group names
}
If combined is TRUE, output is a list of data.frame objects of length equal to the number of groups,
indicating the combined log p-values, log FDR values and log fold changes.
}
\description{
Functions adapted from the deprecated scran::findMarkers to identify the most differential genes through
pairwise t-tests with p-values combined via various methods.
}
\examples{
set.seed(1000)
sce <- scuttle::mockSCE()
sce <- scrapper::normalizeRnaCounts.se(sce)
rownames(sce) <- paste0('Gene_',1:2000)
simple_output <- findMarkersSimple(sce, sce$Treatment, combined=FALSE, 
    assay.type="logcounts", direction="any")
simple_output_combined <- findMarkersSimple(sce, sce$Treatment, combined=TRUE, 
    assay.type="logcounts", direction="any")
}
