% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_bed}
\alias{human_bed}
\title{Datasets containing the genomic coordinates of human TF binding sites}
\format{
A list of GRanges objects
\describe{
\item{seqnames}{chromosome}
\item{ranges}{genomic coordinates}
\item{strand}{strand information}
...
}
}
\source{
\url{http://http://cistrome.org/} \url{https://www.encodeproject.org/}
}
\usage{
human_bed
}
\description{
A list of GRanges objects containing the genomic coordinates of human TF binding sites
curated from Cistrome and ENCODE databases
}
\keyword{datasets}
