% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\docType{data}
\name{human_pwms_v1}
\alias{human_pwms_v1}
\title{human_pwms_v1}
\description{
human_pwms_v1
}
\details{
These motifs were curated from the cisBP database. Position
frequency matrices were converted to PWMs by taking the log
of the frequencies (after adding a pseudocount of 0.008) divided by 0.25.
}
\keyword{datasets}
