% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiregulon_igraph.R
\name{normalize_centrality}
\alias{normalize_centrality}
\title{Normalize centrality scores}
\usage{
normalize_centrality(graph, FUN = identity, weighted = TRUE)
}
\arguments{
\item{graph}{nn igraph object. In a typical use case it will be the output of
\code{add_centrality_degree} function}

\item{FUN}{a function which will be used for normalization. The input to this
function will be the number of edges connected with each node (incident edges).}

\item{weighted}{a logical indicating whether a weighted graph is used}
}
\value{
An igraph object with normalized values of 'centrality' attribute
}
\description{
The nornalized values of centrality will be calculated as old value
divided by \code{FUN}(total number of non-zero edges associated with each node).
}
