% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiregulon_igraph.R
\name{plot_difference_network}
\alias{plot_difference_network}
\title{Plot graph according to grouping factor}
\usage{
plot_difference_network(
  regulon,
  cutoff = 0.01,
  tf = NULL,
  groups = NULL,
  layout = "stress"
)
}
\arguments{
\item{regulon}{an object returned by the getRegulon or addWeights function}

\item{cutoff}{a numerical used to select values of the variables passed in \code{groups}
parameter. Values greater than \code{cutoff} are retained and used as
graph edge weights.}

\item{tf}{a character vector storing the names of transcription factors to be
included in the graph}

\item{groups}{a character indicating levels of the grouping factor; should
correspond to columns of \code{regulon} to be used as \code{weight}
variable in the long format of \code{epiregulon} data frame.}

\item{layout}{a layout specification. Any values that are valid for
\link[ggraph]{ggraph} or \link[ggraph]{create_layout} will work.}
}
\value{
a ggraph object
}
\description{
Plot graph with separate weights for different levels of the grouping factor
}
\examples{
#' # create an artificial getRegulon output
set.seed(1234)
tf_set <- apply(expand.grid(LETTERS[1:10], LETTERS[1:10]),1,  paste, collapse = "")
regulon <- data.frame(tf = sample(tf_set, 5e3, replace = TRUE))
gene_set <- expand.grid(LETTERS[1:10], LETTERS[1:10], LETTERS[1:10])
gene_set <- apply(gene_set,1,function(x) paste0(x,collapse=""))
regulon$target <- sample(gene_set, 5e3, replace = TRUE)
regulon$idxATAC <- 1:5e3
regulon <- cbind(regulon, data.frame(C1 = runif(5e3), C2 = runif(5e3),
C3 = runif(5e3)))
plot_difference_network(regulon, tf = unique(tf_set)[1:3],
groups = c("C1", "C2", "C3"), cutoff = 0.2)
}
\author{
Xiaodai Yao, Tomasz Wlodarczyk
}
