% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotActivityDim}
\alias{plotActivityDim}
\title{A function to plot cell-level reduced dimension results stored in a SingleCellExperiment object, colored by activities for a list of TFs}
\usage{
plotActivityDim(
  sce,
  activity_matrix,
  tf,
  dimtype = "UMAP",
  label = NULL,
  ncol = NULL,
  combine = TRUE,
  legend.label = "activity",
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing dimensionality reduction coordinates}

\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{tf}{A character vector indicating the names of the transcription factors to be plotted}

\item{dimtype}{String indicating the name of dimensionality reduction matrix to be extracted from the SingleCellExperiment}

\item{label}{logical to determine whether the cluster/group labels should be annotated on plot}

\item{ncol}{A integer to specify the number of columns in the combined plot, if combine == TRUE}

\item{combine}{logical to indicate whether to combine and visualize the plots in one panel}

\item{legend.label}{String indicating legend label}

\item{...}{Additional arguments from scater::plotReducedDim}
}
\value{
A combined ggplot object or a list of ggplots if combine == FALSE
}
\description{
A function to plot cell-level reduced dimension results stored in a SingleCellExperiment object, colored by activities for a list of TFs
}
\examples{
\dontrun{
plotActivityDim(sce, score.combine, c("FOXA1","GATA3","SOX9","SPI1"), "TSNE", point_size = 0.25)
}


}
