% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotActivityDim_}
\alias{plotActivityDim_}
\title{A function to plot cell-level reduced dimension results stored in a SingleCellExperiment object, colored by activities for a specific TF}
\usage{
plotActivityDim_(sce, activity_matrix, tf, dimtype, label, legend.label, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing dimensionality reduction coordinates}

\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{tf}{A character vector indicating the names of the transcription factors to be plotted}

\item{dimtype}{String indicating the name of dimensionality reduction matrix to be extracted from the SingleCellExperiment}

\item{label}{A Boolean value to determine whether the cluser/group labels should be annotated on plot}

\item{legend.label}{String indicating legend label}
}
\value{
A ggplot object
}
\description{
A function to plot cell-level reduced dimension results stored in a SingleCellExperiment object, colored by activities for a specific TF
}
