% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotActivityViolin_}
\alias{plotActivityViolin_}
\title{Generate violin plots of inferred activities for a specific TF grouped by cluster/group labels}
\usage{
plotActivityViolin_(activity_matrix, tf, class, legend.label)
}
\arguments{
\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{tf}{A character vector indicating the names of the transcription factors to be plotted}

\item{class}{A character or integer vector of cluster or group labels for single cells}

\item{legend.label}{String indicating the name of variable to be plotted on the legend}
}
\value{
A ggplot object
}
\description{
Generate violin plots of inferred activities for a specific TF grouped by cluster/group labels
}
\author{
Xiaosai Yao, Shang-yang Chen
}
