% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{plotGseaNetwork}
\alias{plotGseaNetwork}
\title{plot networks graph of genesets from regulonEnrich results}
\usage{
plotGseaNetwork(
  tf,
  enrichresults,
  ntop_pathways = 10,
  p.adj_cutoff = 0.05,
  layout = layout.fruchterman.reingold
)
}
\arguments{
\item{tf}{A vector of gene names to be plotted. They should be present in enrichresults}

\item{enrichresults}{Output from regulonEnrich that computes enriched genesets from user-specified regulons
of interest}

\item{ntop_pathways}{An integer indicating the number of top pathways to be included in the graph}

\item{p.adj_cutoff}{p.adjusted cutoff for pathways to be included in the graph. Default value is 0.05}

\item{layout}{layout option from igraph}
}
\value{
an igraph plot of interconnected pathways through TFs
}
\description{
plot networks graph of genesets from regulonEnrich results
}
\examples{
\dontrun{plotGseaNetwork(tf = names(enrichresults), enrichresults, p.adj_cutoff = 0.1,
ntop_pathways = 10)}
}
