% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiregulon_igraph.R
\name{rank_tfs}
\alias{rank_tfs}
\title{Rank transcription factors}
\usage{
rank_tfs(graph, type_attr = "type")
}
\arguments{
\item{graph}{an igraph object with \code{centrality} attribute added to vertices}
}
\value{
a data.frame with transcription factors sorted according to the value of the
\code{centrality} attribute
}
\description{
Rank transcription factors according to degree centrality of their vertices
}
