% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalTests.R
\name{regulonEnrich}
\alias{regulonEnrich}
\title{A function to perform geneset enrichment of user-defined regulons}
\usage{
regulonEnrich(TF, regulon, corr = "weight", corr_cutoff = 0.5, genesets)
}
\arguments{
\item{TF}{A vector of TF names}

\item{regulon}{a weighted regulon consisting of tf, targets, corr and weight}

\item{corr}{name of column that is used to filter target genes for geneset enrichment. Default is "weight".}

\item{corr_cutoff}{a cutoff to filter on the column specified by corr. Default is 0.5.}

\item{genesets}{a dataframe with the first column being the name of the geneset and the second column being the name of the genes}
}
\value{
a combined ggplot object or a list of ggplots if combine == FALSE
}
\description{
A function to perform geneset enrichment of user-defined regulons
}
\examples{
#retrieve genesets
H <- EnrichmentBrowser::getGenesets(org = "hsa", db = "msigdb", cat = "H", gene.id.type = "SYMBOL" )
C6 <- EnrichmentBrowser::getGenesets(org = "hsa", db = "msigdb", cat = "C6", gene.id.type = "SYMBOL" )
#combine genesets and convert genesets to be compatible with enricher
gs <- c(H,C6)
gs.list <- do.call(rbind,lapply(names(gs), function(x) {data.frame(gs=x, genes=gs[[x]])}))

> head(gs.list)
gs   genes
1 M5890_HALLMARK_TNFA_SIGNALING_VIA_NFKB   ABCA1
2 M5890_HALLMARK_TNFA_SIGNALING_VIA_NFKB   ACKR3
3 M5890_HALLMARK_TNFA_SIGNALING_VIA_NFKB    AREG

enrichment_results <- regulonEnrich(c("NKX2-1","GATA6","AR"), regulon=regulon.w, genesets=gs.list)

}
