% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulonPruning.R
\name{addLogFC}
\alias{addLogFC}
\title{Add log fold changes of gene expression to regulons}
\usage{
addLogFC(
  expMatrix,
  clusters,
  regulon,
  direction = c("any", "up", "down"),
  pval.type = c("any", "some", "all"),
  sig_type = c("log.FDR", "log.p.value"),
  logFC_condition = NULL,
  logFC_ref = NULL,
  min.prop = NULL,
  assay.type = "logcounts"
)
}
\arguments{
\item{expMatrix}{A SingleCellExperiment object or matrix containing gene expression with
genes in the rows and cells in the columns. Gene expression should be in logcounts}

\item{clusters}{A character or integer vector of cluster or group labels for single cells}

\item{regulon}{A data frame informing the gene regulatory relationship with the \code{tf} column
representing transcription factors, \code{idxATAC} corresponding to the index in the peakMatrix and
\code{target} column representing target genes}

\item{direction}{tring specifying the direction of log-fold changes to be considered in the alternative hypothesis.}

\item{pval.type}{String specifying how p-values are to be combined across pairwise comparisons for a given group/cluster.}

\item{sig_type}{String specifying whether to use "FDR" or "p.value" for sig_cutoff}

\item{logFC_condition}{A scalar or vector of string indicating the sample names to be compared against \code{logFC_ref}}

\item{logFC_ref}{A scalar indicating the reference sample used to compute logFC. Default value is
\code{rest} which is an average of all pairwise comparisons. Users can also specify a reference sample, for example, \code{DMSO}.}

\item{min.prop}{Numeric scalar specifying the minimum proportion of significant comparisons per gene.
Defaults to 0.5 when pval.type="some", otherwise defaults to zero.}

\item{assay.type}{A string specifying which assay values to use, usually "logcounts".}
}
\value{
A DataFrame of regulons with additional columns of logFC and significance
}
\description{
Add log fold changes of gene expression to regulons
}
\details{
The choice of \code{pval.type} determines whether the highly ranked genes are those that are DE between the current group and:
\itemize{
\item any other group (\code{"any"})
\item all other groups (\code{"all"})
\item some other groups (\code{"some"})
}
Adapted from the deprecated scran::findMarkers function
}
\examples{
# create a mock SingleCellExperiment object for gene expression matrix
set.seed(1000)
gene_sce <- scuttle::mockSCE()
gene_sce <- scrapper::normalizeRnaCounts.se(gene_sce)
rownames(gene_sce) <- paste0('Gene_',1:2000)

# create a mock regulon
regulon <- data.frame(tf=c(rep('Gene_1',10), rep('Gene_2',10)),
    idxATAC=sample(1:100, 20),
    target=c(paste0('Gene_', sample(3:2000,10)),
    paste0('Gene_',sample(3:2000,10))))

# filter regulon
pruned.regulon <- addLogFC(expMatrix=gene_sce, 
    clusters=gene_sce$Treatment,
    regulon=regulon,
    sig_type="log.p.value")

}
\author{
Xiaosai Yao
}
