% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRegulon.R
\name{addTFMotifInfo}
\alias{addTFMotifInfo}
\title{A function to add TF binding motif occupancy information to the peak2gene object}
\usage{
addTFMotifInfo(p2g, grl, peakMatrix = NULL, archR_project_path = NULL)
}
\arguments{
\item{p2g}{A Peak2Gene dataframe created by ArchR or getP2Glinks() function}

\item{grl}{GRangeList object containing reference TF binding information}

\item{peakMatrix}{A matrix of scATAC-seq peak regions with peak ids as rows}

\item{archR_project_path}{Path to a ArchR project that have performed LSI dimensionality reduction and scRNA-seq integration}
}
\value{
A dataframe containing overlapping ids of scATAC-seq peak regions and reference TF binding regions
}
\description{
A function to add TF binding motif occupancy information to the peak2gene object
}
\examples{
\dontrun{
p2g <- getP2Glinks("/gstore/project/lineage/sam/heme_GRN/OUTPUT")
grl <- getTFMotifInfo(genome="mm10")
overlap <- addTFMotifInfo(p2g, grl, archR_project_path = "/gstore/project/lineage/sam/heme_GRN/OUTPUT")
head(overlap)
}
}
