% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRegulon.R
\name{addTFMotifInfo}
\alias{addTFMotifInfo}
\title{A function to add TF binding motif occupancy information to the peak2gene object}
\usage{
addTFMotifInfo(archr_path, organism = "human")
}
\arguments{
\item{archr_path}{Path to a ArchR project that have performed LSI dimensionality reduction and scRNA-seq integration}

\item{organism}{human or mouse are currently supported}
}
\value{
None, a motif binary matrix file will be created in the Annotations folder of supplied ArchR project path as ChIP-Matches-In-Peaks.rds
}
\description{
A function to add TF binding motif occupancy information to the peak2gene object
}
\examples{
1+1
}
