% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWeights.R
\name{addWeights}
\alias{addWeights}
\title{A function to calculate weights for regulons}
\usage{
addWeights(
  regulon,
  sce,
  cluster_factor,
  block_factor = NULL,
  exprs_values = NULL,
  corr = TRUE,
  MI = TRUE,
  multicore = TRUE
)
}
\arguments{
\item{regulon}{a regulon in the form of a tall format matrix consisting of tf(regulator), target and a column indicating degree of association between TF and target such as "mor" or "corr".
example regulon:
tf      target  corr}

\item{sce}{scater single cell object compute weights on}

\item{cluster_factor}{cluster labels of the scater single cell object}

\item{block_factor}{batch labels of the scater single cell object}

\item{exprs_values}{name of the assay to be retrieved from scater object}

\item{corr}{whether to calculate weights based on correlation}

\item{MI}{whether to calculate weights based on mutual information}

\item{multicore}{number of cores to use}
}
\value{
dataframe with three columns in the form of regulon input. The third column is replaced with the inferred weights
}
\description{
A function to calculate weights for regulons
}
\examples{
1+1
}
