% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWeights.R
\name{addWeights}
\alias{addWeights}
\title{Calculate weights for the regulons by computing co-association between TF and target gene expression}
\usage{
addWeights(
  regulon,
  sce,
  peakMatrix = NULL,
  cluster_factor = NULL,
  block_factor = NULL,
  exprs_values = "logcounts",
  method = "corr",
  MI = FALSE,
  peak_assay = "PeakMatrix",
  min_targets = 10,
  expr_cutoff = 1,
  peak_cutoff = 0,
  BPPARAM = BiocParallel::SerialParam(),
  alt.exp = NULL,
  alt.exp.merge = FALSE
)
}
\arguments{
\item{regulon}{A data frame consisting of tf (regulator) and target in the column names. Additional columns indicating degree
of association between tf and target such as "mor" or "corr" are optional.}

\item{sce}{A SingleCellExperiment object containing gene expression information}

\item{peakMatrix}{A SingleCellExperiment object or matrix containing peak accessibility with
peaks in the rows and cells in the columns}

\item{cluster_factor}{String specifying the field in the colData of the SingleCellExperiment object to be averaged as pseudobulk (such as cluster)}

\item{block_factor}{String specifying the field in the colData of the SingleCellExperiment object to be used as blocking factor (such as batch)}

\item{exprs_values}{String specifying the name of the assay to be retrieved from the SingleCellExperiment object}

\item{method}{String specifying the method of weights calculation. Three options are available: \code{corr}, \code{wilcoxon} and \code{logFC}.}

\item{MI}{Logical scalar indicating whether to calculate weights based on mutual information}

\item{peak_assay}{String indicating the name of the assay in peakMatrix for chromatin accessibility}

\item{min_targets}{Integer specifying the minimum number of targets for each tf in the regulon with 10 targets as the default}

\item{expr_cutoff}{A scalar indicating the minimum gene expression for transcription factor above which
cell is considered as having expressed transcription factor.}

\item{peak_cutoff}{A scalar indicating the minimum peak accessibility above which peak is
considered open.}

\item{BPPARAM}{A BiocParallelParam object specifying whether summation should be parallelized. Use BiocParallel::SerialParam() for
serial evaluation and use BiocParallel::MulticoreParam() for parallel evaluation}

\item{alt.exp}{A matrix that is used in place of gene expression to correlate with target gene expression. See details.}

\item{alt.exp.merge}{A logical to indicate whether to consider both TF expression and alt.exp matrix. See details.}
}
\value{
A data frame with columns of corr and/or MI added to the regulon. TFs not found in the expression matrix and regulons not
meeting the minimal number of targets were filtered out.
}
\description{
Calculate weights for the regulons by computing co-association between TF and target gene expression
}
\details{
The default mode is to compute weights by correlating the pseudobulk target gene expression vs the pseudobulk TF gene expression.
However, often times, an inhibitor of TF does not alter the gene expression of the TF. In rare cases, cells may even compensate
by increasing the expression of the TF. In this case, the activity of the TF, if computed by gene expression correlation, may show a
spurious increase. As an alternative to gene expression, we may use accessibility associated with TF, such as those computed by
chromVar. When alt.exp.merge is true, we take the product of the gene expression and the values in the alt.exp matrix.
When \code{method} is set to \code{wilcoxon} or \code{logFC} for each unique pair of transcription factor and target gene the corresponding
data from peakMatrix is collapsed (by cell-wise summing). This data along with data on the expression of transcription factor is
used to label the cells which show both transcription factor expression and regulatory element accessibility. The target gene expression
in the labeled cells is contrasted against the rest of the cells. If \code{wilcoxon} is chosen, the cell groups are compared with Wilcoxon test
and the weight is the effect size calculated as a z-score divided by the square root of the sample size. If \code{logFC} is chosen the
weight is the difference between mean target gene expression in compared cells groups.
}
\examples{
# create a mock singleCellExperiment object for gene expression matrix
example_sce <- scuttle::mockSCE()
example_sce <- scuttle::logNormCounts(example_sce)
example_sce$cluster <- sample(LETTERS[1:5], ncol(example_sce), replace = TRUE)

# create a mock regulon
regulon <- data.frame(tf = c(rep("Gene_0001",5), rep("Gene_0002",10)),
                      target = c(paste0("Gene_000",2:6), paste0("Gene_00",11:20)))

# add weights to regulon
regulon.w <- addWeights(regulon, example_sce, cluster_factor="cluster", exprs_values = "logcounts",
                        min_targets = 5)

# Alternatively, add a matrix of chromVar values in place of TF expression
\dontrun{
# create chromVar values from archR
library(ArchR)
library(parallel)
proj <- addBgdPeaks(proj)
proj <- addDeviationsMatrix(ArchRProj = proj,
peakAnnotation = "motif", force = TRUE,logFile = "addDeviation")

#retrieve chromVar matrix
#chromVar
motifMatrix <- getMatrixFromProject(
ArchRProj = proj,
useMatrix = "motifMatrix",
useSeqnames = NULL,
verbose = TRUE,
binarize = FALSE,
threads = getArchRThreads(),
logFile = createLogFile("getMatrixFromProject")
)

# calculate weights using alt.exp
regulon.w.2 <- addWeights(regulon, example_sce, cluster_factor = "cluster",
exprs_values = "logcounts", min_targets = 5, alt.exp = assay(motifMatrix, "z"),
alt.exp.merge = TRUE)
}
}
\author{
Xiaosai Yao, Shang-yang Chen, Tomasz Wlodarczyk
}
