% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ArchRMatrix2SCE}
\alias{ArchRMatrix2SCE}
\title{Convert ArchR Gene Expression Matrix to SingleCellExperiment}
\usage{
ArchRMatrix2SCE(rse)
}
\arguments{
\item{rse}{a RangedSummarizedExperiment object obtained from \code{ArchR::getMatrixFromProject}}
}
\value{
A SingleCellExperiment object
}
\description{
Convert ArchR Gene Expression Matrix to SingleCellExperiment
}
\details{
When gene expression matrices (either GeneIntegrationMatrix or GeneExpressionMatrix)
are retrieved from ArchR project, they are returned as \code{RangedSummarizedExperiment}. This function
converts the class of the object from \code{RangedSummarizedExperiment} to \code{SingleCellExperiment}.
During the conversion, the genomic location of the genes are transferred from \code{rowData} in the
\code{RangedSummarizedExperiment} to the \code{rowRanges} of the \code{SingleCellExperiment}.
}
