% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateActivity.R
\name{calculateActivity}
\alias{calculateActivity}
\title{A function that calculates the per cell activity of master regulators based on a regulon}
\usage{
calculateActivity(
  sce,
  regulon,
  mode = "weight",
  method = "weightedmean",
  ncore = 1,
  assay = "logcounts"
)
}
\arguments{
\item{sce}{A SingleCellExperiment object with rows representing genes and columns represent cells. Rownames (either gene symbols or geneID) must be consistent with the naming convention in the regulon.}

\item{regulon}{A data frame consisting of tf (regulator) and target in the column names, with additional columns indicating degree of association between tf and target such as "mor" or "corr" obtained from addWeights.}

\item{mode}{String indicating the name of column to be used as the weights}

\item{method}{String indicating the method for calculating activity. Available methods are weightedMean or aucell}

\item{ncore}{Integer specifying the number of cores to be used in AUCell}

\item{assay}{String specifying the name of the assay to be retrieved from the SingleCellExperiment object. Set to "logcounts" as the default}
}
\value{
a matrix of inferred transcription factor (row) activities in single cells (columns)
}
\description{
A function that calculates the per cell activity of master regulators based on a regulon
}
\examples{
example_sce <- scuttle::mockSCE()
example_sce <- scuttle::logNormCounts(example_sce)
example_sce$cluster <- sample(LETTERS[1:5], ncol(example_sce), replace = TRUE)
regulon <- data.frame(tf = c(rep("Gene_0001",5), rep("Gene_0002",10)),
 target = c(paste0("Gene_000",2:6), paste0("Gene_00",11:20)))
regulon.w <- addWeights(regulon, example_sce, cluster_factor = "cluster",
 exprs_values = "counts", min_targets = 5)
activity <- calculateActivity(example_sce, regulon.w, assay = "logcounts")
}
