% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateActivity.R
\name{calculateActivity}
\alias{calculateActivity}
\title{A function that calculates the per cell activity of master regulons based on a regulon}
\usage{
calculateActivity(sce, regulon, mode, method = NULL, ncore = NULL, assay)
}
\arguments{
\item{sce}{a matrix of genes by cells. Rows represent genes and columns represent cells. Rownames (either gene symbols or geneID) must be consistent with the naming convention in the regulon.}

\item{regulon}{a regulon in the form of a tall format matrix consisting of tf(regulator), target and a column indicating degree of association between TF and target such as "mor" or "corr".
example regulon:
tf      target  corr
Esr1    Pgr     0.56}

\item{mode}{a string indicating the mode of regulon such as "mor" or "corr" when weightedMean is the chosen method}

\item{method}{method for calculating activity. Available methods are weightedMean or aucell. The parameter used for the weights in weightedMean is further specified by mode.}

\item{ncore}{number of cores to use}
}
\value{
a matrix of inferred transcription factor (row) activities in single cells (columns)
}
\description{
A function that calculates the per cell activity of master regulons based on a regulon
}
\examples{
1+1
}
