% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateActivityBulk.R
\name{calculateActivityBulk}
\alias{calculateActivityBulk}
\title{A function to calculate transcriptional activity from a matrix of gene expression and regulon}
\usage{
calculateActivityBulk(
  expr,
  regulon,
  method = c("gsva", "ssgsea", "zscore", "plage"),
  mode = c("mor", "corr"),
  corr_calculate = NULL,
  save.regulon = NULL,
  sample_n = NULL,
  assay_name = "rpkm",
  rowData_name = "symbol"
)
}
\arguments{
\item{expr}{Matrix of gene expression with gene names matching those of the regulon or a SummarizedExperiment object}

\item{regulon}{Matrix consisting of tf(regulator), target and a column indicating degree of association between TF and target}

\item{method}{String indicating the method used to calculate signature score. Available methods in gsva package include "gsva", "ssgsea", "zscore" and "plage"}

\item{mode}{String indicating the choice of TF-target association being either correlation "corr" or mode of regulation "mor".
Correlation can either be supplied or computed from expr using (corr_calculate = FALSE). Mor is supplied by the Dorothea regulon.}

\item{corr_calculate}{Logical indicating whether correlation needs to be calculated}

\item{save.regulon}{String indicating the path to save the new regulon with updated correlation}

\item{sample_n}{An integer indicating the number of samples to subsample if the original expression matrix is large}

\item{assay_name}{String indicating the name of the assay corresponding to gene expression
if expr is a SummarizedExperiment object}

\item{rowData_name}{String indicating the column name in the rowData that matches the gene names in the regulon
if expr is a SummarizedExperiment object}
}
\value{
A matrix of inferred transcription factor (row) activities in samples (columns)
}
\description{
This function generates the transcriptional activity from a matrix of bulk gene expression data and pre-existing regulons such as those
available from Dorothea. The direction of target genes (activating or repressive) can be defined either by the mode of regulation
included in the regulon or calculated based on the correlation between the expression of the TF and its target genes
}
\examples{
# Load regulon from dorothea
library(dorothea)
data(dorothea_hs, package = "dorothea")
regulon <- dorothea_hs[which(dorothea_hs$confidence \%in\% c("A")),]

# Create a mock expression matrix
genes <- unique(c(regulon$tf, regulon$target))
expr <- matrix(rnorm(length(genes)*20, mean=2), nrow = length(genes), ncol=20)
rownames(expr) <- genes
colnames(expr) <- paste("sample", 1:20)

# Use mor embedded in Dorothea regulon
activity.mor <- calculateActivityBulk(expr, regulon, mode = "mor", method = "ssgsea")

# Add correlation of Dorothea from user supplied gene matrix
activity.corr <- calculateActivityBulk(expr, regulon, mode = "corr",
method ="ssgsea", corr_calculate = TRUE)

# Subsample
activity.corr <- calculateActivityBulk(expr, regulon, mode = "corr",
method ="ssgsea", corr_calculate = TRUE, sample_n = 10)

# Use Dorothea regulon with pre-calculated calculation
\dontrun{
regulon <- readRDS(file = "/gstore/project/lineage/GRN/data/Dorothea_TCGA/regulon.Breast.TCGA.ABCDE.rds")
activity.corr <- calculateActivityBulk(expr, regulon, mode = "corr", method ="ssgsea")
}

}
