% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateP2G.R
\name{calculateP2G}
\alias{calculateP2G}
\title{A function to establish peak to gene links based on correlations between ATAC-seq peaks and RNA-seq genes}
\usage{
calculateP2G(
  peakMatrix = NULL,
  expMatrix = NULL,
  reducedDim = NULL,
  ArchR_path = NULL,
  reducedDimName = "LSI",
  useDim = "IterativeLSI",
  useMatrix = "GeneIntegrationMatrix",
  cor_cutoff = 0.5,
  cellNum = 200,
  seed = 1,
  ...
)
}
\arguments{
\item{peakMatrix}{A SingleCellExperiment object containing counts of chromatin accessibility at each peak region or genomic bin from scATAC-seq}

\item{expMatrix}{A SingleCellExperiment object containing gene expression counts from scRNA-seq}

\item{reducedDim}{A matrix of dimension reduced values, for example derived from IterativeLSI algorithm of ArchR}

\item{ArchR_path}{String specifying the path to an ArchR project if ArchR's implementation of addPeak2GeneLinks is desired}

\item{reducedDimName}{String specifying the name of the reduced dimension matrix}

\item{useDim}{String specifying which dimensional reduction representation in the ArchR project to use}

\item{useMatrix}{String specifying which data matrix in the ArchR project to use}

\item{cor_cutoff}{A numeric scalar to specify the correlation cutoff between ATAC-seq peaks and RNA-seq genes to assign peak to gene links. Default correlation cutoff is 0.5.}

\item{cellNum}{An integer to specify the number of cells to include in each K-means cluster. Default is 200 cells.}

\item{seed}{An integer scalar to specify the seed for K-means clustering}

\item{...}{other parameters to pass to addPeak2GeneLinks from ArchR package}
}
\value{
A Peak2Gene correlation datafrane
}
\description{
A function to establish peak to gene links based on correlations between ATAC-seq peaks and RNA-seq genes
}
\examples{
\dontrun{
p2g <- calculateP2G(peakmatrix, expmatrix, reducedDim)
head(p2g)
}
}
