% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateP2G.R
\name{calculateP2G}
\alias{calculateP2G}
\title{A function to compute correlations between ATAC-seq peaks and RNA-seq genes}
\usage{
calculateP2G(
  peakMatrix = NULL,
  expMatrix = NULL,
  reducedDim = NULL,
  ArchR_path = NULL,
  useDim = "IterativeLSI",
  useMatrix = "GeneIntegrationMatrix",
  cor_cutoff = 0.5,
  cellNum = 200,
  seed = 1,
  ...
)
}
\arguments{
\item{peakMatrix}{SingleCellExperiment object containing counts of scATAC-seq peaks}

\item{expMatrix}{SingleCellExperiment object containing counts of scRNA-seq expressions}

\item{reducedDim}{matrix of dimensional reduced values, can be derived from IterativeLSI algorithm of ArchR}

\item{ArchR_path}{string specifying optional path to a ArchR project if ArchR's implementation of addPeak2GeneLinks is desired}

\item{useDim}{String specifying which dimensional reduction representation in the ArchR project to use}

\item{useMatrix}{String specifying which data matrix in the ArchR project to use}

\item{cor_cutoff}{cutoff for correlations between ATAC-seq peaks and RNA-seq genes}

\item{cellNum}{number of cells to include in each K-means cluster, to control number of k}

\item{seed}{number of random seed to use for K-means clustering}

\item{...}{other parameters to pass to addPeak2GeneLinks from ArchR package}
}
\value{
A Peak2Gene correlation datafrane
}
\description{
A function to compute correlations between ATAC-seq peaks and RNA-seq genes
}
\examples{
p2g <- calculateP2G(peakmatrix, expmatrix, reducedDim)
head(p2g)

}
