% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{enrichPlot}
\alias{enrichPlot}
\title{A function to plot results of regulonEnrich}
\usage{
enrichPlot(results, top = 15, ncol = 3, combine = TRUE)
}
\arguments{
\item{results}{Output from regulonEnrich}

\item{top}{An integer to indicate the number of pathways to plot ranked by significance. Default is 15.}

\item{ncol}{An integer to indciate the number of columns in the combined plot, if combine == TRUE. Default is 3.}

\item{combine}{logical to indicate whether to combine and visualize the plots in one panel}
}
\value{
A combined ggplot object or a list of ggplots if combine == FALSE
}
\description{
A function to plot results of regulonEnrich
}
\examples{
\dontrun{
enrichplot(results=enrichment_results )
}

}
