% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalTests.R
\name{findDifferentialActivity}
\alias{findDifferentialActivity}
\title{A function to test for differential TF activity between pairs of single cell clusters/groups}
\usage{
findDifferentialActivity(
  activity_matrix,
  groups,
  test.type = "t",
  pval.type = "some",
  direction = "up",
  ...
)
}
\arguments{
\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{groups}{A character or integer vector of cluster or group labels for single cells}

\item{test.type}{String indicating the type of statistical tests to be passed to scran::findMarkers, can be "t", "wilcox". or "binom"}

\item{pval.type}{A string specifying how p-values are to be combined across pairwise comparisons for a given group/cluster.}

\item{direction}{A string specifying direction of differential TF activity, can be "up" or "down"}

\item{...}{Further arguments to pass to scran::findMarkers}
}
\value{
A named list of dataframes containing differential TF activity test results for each cluster/group
}
\description{
A function to test for differential TF activity between pairs of single cell clusters/groups
}
\examples{
\dontrun{
# score.combine is TF activity score matrix from calculateActivity function
markers <- findDifferentialActivity(score.combine, sce$BioClassification, pval.type="some", direction="up", test.type= "t")
}
}
