% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRegulon.R
\name{getP2Glinks}
\alias{getP2Glinks}
\title{A function to compute correlations between ATAC-seq peaks and RNA-seq genes}
\usage{
getP2Glinks(
  archr_path,
  cor_cutoff = 0.5,
  reducedDims = "IterativeLSI",
  useMatrix = "GeneIntegrationMatrix",
  ...
)
}
\arguments{
\item{archr_path}{Path to a ArchR project that have performed LSI dimensionality reduction and scRNA-seq integration}

\item{cor_cutoff}{Cutoff for correlations between ATAC-seq peaks and RNA-seq genes}
}
\value{
A Peak2Gene object (DFrame)
}
\description{
A function to compute correlations between ATAC-seq peaks and RNA-seq genes
}
\examples{
1+1

}
