% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRegulon.R
\name{getRegulon}
\alias{getRegulon}
\title{A function to combine the TF binding motif info and peak to gene correlations to generate regulons}
\usage{
getRegulon(archr_path, p2g_object, aggregate = TRUE)
}
\arguments{
\item{archr_path}{Path to a ArchR project that have performed LSI dimensionality reduction and scRNA-seq integration}

\item{p2g_object}{A Peak2Gene object (DFrame) created by ArchR or getP2Glinks() function}

\item{aggregate}{boolean value that specify whether peak and gene ids are kept in regulon output or not}
}
\value{
a tall format dataframe consisting of tf(regulator), target and a column indicating degree of association between TF and target such as "mor" or "corr".
example regulon:
tf      target  corr
}
\description{
A function to combine the TF binding motif info and peak to gene correlations to generate regulons
}
\examples{
1+1
}
