% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRegulon.R
\name{getRegulon}
\alias{getRegulon}
\title{A function to combine the TF binding motif info and peak to gene correlations to generate regulons}
\usage{
getRegulon(p2g, overlap, aggregate = TRUE)
}
\arguments{
\item{p2g}{A Peak2Gene dataframe created by ArchR or getP2Glinks() function}

\item{overlap}{dataframe storing overlaps between the regions of the peak matrix with the bulk TF ChIP-seq binding sites}

\item{aggregate}{boolean value that specify whether peak and gene ids are kept in regulon output or not}
}
\value{
a tall format dataframe consisting of tf(regulator), target and a column indicating degree of association between TF and target such as "mor" or "corr".
example regulon:
tf      target  corr
}
\description{
A function to combine the TF binding motif info and peak to gene correlations to generate regulons
}
\examples{
p2g <- getP2Glinks("/gstore/project/lineage/sam/heme_GRN/OUTPUT")
grl <- getTFMotifInfo(genome="mm10")
overlap <- addTFMotifInfo(p2g, grl, archR_project_path = "/gstore/project/lineage/sam/heme_GRN/OUTPUT")
regulon <- getRegulon(p2g, overlap, aggregate = T)
head(regulon)
}
