% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalTests.R
\name{getSigGenes}
\alias{getSigGenes}
\title{A function to compile and summarize the output from findDifferentialActivity function}
\usage{
getSigGenes(
  da_list,
  fdr_cutoff = 0.05,
  logFC_cutoff = NULL,
  order = "FDR",
  topgenes = NULL,
  decreasing = FALSE
)
}
\arguments{
\item{da_list}{List of dataframes from running findDifferentialActivity}

\item{fdr_cutoff}{cutoff for FDR value, default is 0.05}

\item{logFC_cutoff}{cutoff for log fold change}

\item{order}{variable to order the differential TF results by, default is FDR}

\item{topgenes}{number of top ordered genes to include in output}

\item{decreasing}{direction of ordering for differential TF activity results, default is FALSE}
}
\value{
A compiled dataframe of TFs with differential activities across clusters/groups
}
\description{
A function to compile and summarize the output from findDifferentialActivity function
}
\examples{
1+1 = 2
}
