% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRegulon.R
\name{getTFMotifInfo}
\alias{getTFMotifInfo}
\title{An accessor function to retrieve TF motif info from genomitory repository}
\usage{
getTFMotifInfo(genome = "hg19")
}
\arguments{
\item{genome}{a string specifying the genome for bed files}
}
\value{
A GRangeList object containing binding site information of 1274 TFs
}
\description{
An accessor function to retrieve TF motif info from genomitory repository
}
\examples{
grl <- getTFMotifInfo(genome="mm10")
head(grl)

}
