% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRegulon.R
\name{getTFMotifInfo}
\alias{getTFMotifInfo}
\title{An accessor function to retrieve TF motif info from genomitory repository}
\usage{
getTFMotifInfo()
}
\value{
A GRangeList object containing binding site information of 1274 TFs
}
\description{
An accessor function to retrieve TF motif info from genomitory repository
}
\examples{
1+1
}
