% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRegulon.R
\name{getTFMotifInfo}
\alias{getTFMotifInfo}
\title{Retrieve TF binding sites or motif positions}
\usage{
getTFMotifInfo(
  genome = c("hg38", "hg19", "mm10"),
  source = c("atlas", "cistrome", "encode.sample", "atlas.sample", "atlas.tissue"),
  metadata = FALSE,
  mode = c("occupancy", "motif"),
  peaks = NULL
)
}
\arguments{
\item{mode}{a string indicating whether to download a GRangelist of TF binding sites ('occupancy') or motif matches ('motif').
TF binding information is retrieved from  \link[scMultiome:tfBinding]{scMultiome::tfBinding}. The
motif information was obtained from \href{https://github.com/GreenleafLab/chromVARmotifs}{chromVARmotifs} (human_pwms_v2 and mouse_pwms_v2,
version 0.2 with filtering of cisBP motifs) and is also hosted on scMultiome.}

\item{peaks}{A GRanges object indicating the peaks to perform motif annotation on.
The peak indices should match the \code{idxATAC} column in the regulon.}
}
\description{
Combined transcription factor ChIP-seq data from ChIP-Atlas and ENCODE or
from CistromeDB and ENCODE.
}
\examples{
# retrieve TF binding info
\donttest{
getTFMotifInfo('mm10', 'atlas')
}

# retrieve motif info
peaks <- GRanges(seqnames = c('chr12','chr19','chr19','chr11','chr6'),
ranges = IRanges(start = c(124914563,50850845, 50850844, 101034172, 151616327),
end = c(124914662,50850929, 50850929, 101034277, 151616394)))
grl <- getTFMotifInfo(genome = 'hg38', mode = 'motif', peaks=peaks)

}
