% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateActivity.R
\name{pathwayscoreCoeffNorm}
\alias{pathwayscoreCoeffNorm}
\title{A subfunction for calculateActivity for deriving individual}
\usage{
pathwayscoreCoeffNorm(mat_scale, genenames, pathway, tf_name)
}
\arguments{
\item{mat_scale}{log2 normalized expression matrix from single cell data}

\item{genenames}{vector of gene names in mat_scale}

\item{pathway}{a dataframe containing gene names and weights of regulons belong to the TF}

\item{tf_name}{name of the transcription factor being computed scores for}
}
\value{
A dataframe of a single column with inferred activity scores in single cells for the TF
}
\description{
A subfunction for calculateActivity for deriving individual
}
\examples{
1+1
}
