% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotActivityDim}
\alias{plotActivityDim}
\title{A function to plot cell-level reduced dimension results stored in a SingleCellExperiment object, colored by activities for a list of TFs}
\usage{
plotActivityDim(
  sce,
  activity_matrix,
  tf,
  dimtype = "UMAP",
  label = NULL,
  ncol = NULL,
  combine = TRUE,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing dimensional reduction coordinates (UMAP or tSNE)}

\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{tf}{The names of the transcription factors to be plotted}

\item{dimtype}{Name of dimensional reduction to be plotted (UMAP or tSNE)}

\item{label}{A boolean value to determine whether the cluser/group labels should be annotated on plot}

\item{ncol}{number of columns in the combined plot, if combine == TRUE}

\item{combine}{whether to combine and visualize the plots in one panel or not}

\item{...}{}
}
\value{
a combined ggplot object or a list of ggplots if combine == FALSE
}
\description{
A function to plot cell-level reduced dimension results stored in a SingleCellExperiment object, colored by activities for a list of TFs
}
\examples{
1+1 = 2
}
