% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotActivityDim_}
\alias{plotActivityDim_}
\title{Plot cell-level reduced dimension results stored in a SingleCellExperiment object, colored by activities for a specific TF}
\usage{
plotActivityDim_(
  sce,
  activity_matrix,
  tf,
  dimtype,
  label,
  legend.label,
  colors,
  limit,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing dimensionality reduction coordinates}

\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{tf}{A character vector indicating the names of the transcription factors to be plotted}

\item{dimtype}{String indicating the name of dimensionality reduction matrix to be extracted from the SingleCellExperiment}

\item{label}{String corresponding to the field in the colData of sce for annotation on plot}

\item{legend.label}{String indicating the name of variable to be plotted on the legend}

\item{colors}{A vector of 2 colors for the intensity, with the first element refering to the lower value and
the second element refering to the higher value. Default is c("blue","yellow").}

\item{limit}{A vector of lower and upper bounds for the color scale. The default option is NULL and will adjust
to minimal and maximal values}

\item{...}{Additional arguments from scater::plotReducedDim}
}
\value{
A ggplot object
}
\description{
Plot cell-level reduced dimension results stored in a SingleCellExperiment object, colored by activities for a specific TF
}
\author{
Xiaosai Yao, Shang-yang Chen
}
