% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotActivityDim_}
\alias{plotActivityDim_}
\title{A function to plot cell-level reduced dimension results stored in a SingleCellExperiment object, colored by activities for a specific TF}
\usage{
plotActivityDim_(sce, activity_matrix, tf, dimtype, label, legend.label, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing dimensional reduction coordinates (UMAP or tSNE)}

\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{tf}{The name of the transcription factor to be plotted}

\item{dimtype}{Name of dimensional reduction to be plotted (UMAP or tSNE)}

\item{label}{A boolean value to determine whether the cluser/group labels should be annotated on plot}
}
\value{
A ggplot object
}
\description{
A function to plot cell-level reduced dimension results stored in a SingleCellExperiment object, colored by activities for a specific TF
}
