% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotActivityViolin}
\alias{plotActivityViolin}
\title{A function to draw violin plots of inferred activities for a list of TFs grouped by cluster/group labels}
\usage{
plotActivityViolin(activity_matrix, tf, class, ncol = NULL, combine = TRUE)
}
\arguments{
\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{tf}{A character vector indicating the names of the transcription factors to be plotted}

\item{class}{A vector of cluster or group labels for single cells}

\item{ncol}{A integer to indicate the number of columns in the combined plot, if combine == TRUE}

\item{combine}{A Boolean value to indicate whether to combine and visualize the plots in one panel}
}
\value{
A combined ggplot object or a list of ggplots if combine == FALSE
}
\description{
A function to draw violin plots of inferred activities for a list of TFs grouped by cluster/group labels
}
\examples{
\dontrun{
plotActivityViolin(score.combine, c("FOXA1","GATA3","SOX9","SPI1"), sce$BioClassification)
}
}
