% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotActivityViolin_}
\alias{plotActivityViolin_}
\title{A function to draw a violin plot of inferred activities for a specific TF grouped by cluster/group labels}
\usage{
plotActivityViolin_(activity_matrix, tf, class)
}
\arguments{
\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{tf}{The name of the transcription factor to be plotted}

\item{class}{A vector of cluster or group labels for single cells}
}
\value{
A ggplot object
}
\description{
A function to draw a violin plot of inferred activities for a specific TF grouped by cluster/group labels
}
\examples{
1+1 = 2
}
