% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotBubble}
\alias{plotBubble}
\title{A function to draw bubble plot of relative activities across cluster/group labels for a list of TFs}
\usage{
plotBubble(activity_matrix, tf.list, class)
}
\arguments{
\item{activity_matrix}{The name of the transcription factor to be plotted}

\item{tf.list}{The list of the transcription factors to be plotted}

\item{class}{A vector of cluster or group labels for single cells}
}
\value{
A ggplot object
}
\description{
A function to draw bubble plot of relative activities across cluster/group labels for a list of TFs
}
\examples{
1+1 = 2
}
