% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotBubble}
\alias{plotBubble}
\title{A function to draw bubble plot of relative activities across cluster/group labels for a list of TFs}
\usage{
plotBubble(activity_matrix, tf.list, class, bubblesize = "FDR")
}
\arguments{
\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{tf.list}{The list of the transcription factors to be plotted}

\item{class}{A vector of cluster or group labels for single cells}

\item{bubblesize}{A variable from findDifferentialActivity output to scale size of bubbles by, default is FDR}
}
\value{
A ggplot object
}
\description{
A function to draw bubble plot of relative activities across cluster/group labels for a list of TFs
}
\examples{
1+1 = 2
}
