% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotBubble}
\alias{plotBubble}
\title{A function to draw bubble plot of relative activities across cluster/group labels for a list of TFs}
\usage{
plotBubble(activity_matrix, tf, class, bubblesize = "FDR")
}
\arguments{
\item{activity_matrix}{A matrix of TF activities inferred from calculateActivity}

\item{tf}{A character vector indicating the names of the transcription factors to be plotted}

\item{class}{A character or integer vector of cluster or group labels for single cells}

\item{bubblesize}{String indicating the variable from findDifferentialActivity output to scale size of bubbles by. Default is FDR}
}
\value{
A ggplot object
}
\description{
A function to draw bubble plot of relative activities across cluster/group labels for a list of TFs
}
\examples{
\dontrun{
plotBubble(score.combine, c("SPI1", "ZNF623", "IRF4","SOX4"), sce$BioClassification)
}

}
